Add files via upload
Create a Node for a new LinkedList that is capable of representing a Movie. Each Node should
contain the following information: Movie Title, Movie Description, Genre, length in minutes,
and rating (1 - 5 star rating) The Node class should also provide a toString method to produce
the following String representation:
Movie Title
Description: This is a cool movie about trains
Genre: Action
Duration: 120
Rating: 5

Construct a new LinkedList that uses this Node. The LinkedList class should contain the
following:
I. Reference to first and last Nodes – this references should enable for the
identification of the first and last nodes in the linkedlist
II. isEmpty – this method should check if the linkedlist is empty and return a boolean
III. genreSearch - this method should return a String representation of the movies
within the Linkedlist that are identified by the genre
IV. get_average – this method should return the average star rating of all movies stored
within the LinkedList
V. add - this method should allow the addition of a new Node into the linkedlist at the
end of the current list
VI. remove - this method removes the first occurrence of the target Node. The target is
defined by Movie Title. You should take into account when no movie title is found